@@ -11,53 +11,57 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.0.4",
-    "@angular/cli": "^6.0.8",
-    "@angular/common": "^6.0.4",
-    "@angular/compiler": "^6.0.4",
-    "@angular/core": "^6.0.4",
-    "@angular/forms": "^6.0.4",
-    "@angular/http": "^6.0.4",
-    "@angular/platform-browser": "^6.0.4",
-    "@angular/platform-browser-dynamic": "^6.0.4",
-    "@angular/router": "^6.0.4",
-    "@ng-bootstrap/ng-bootstrap": "^2.0.0",
-    "@reactivex/rxjs": "^6.2.0",
-    "@types/moment-timezone": "^0.5.6",
-    "@types/pouchdb": "^6.3.2",
-    "@types/pouchdb-core": "^6.1.13",
-    "@types/uuid": "^3.4.3",
-    "bootstrap": "^4.1.1",
-    "buffer": "^5.1.0",
-    "chart.js": "^2.7.2",
-    "core-js": "^2.5.7",
+    "@angular/animations": "^8.2.14",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/platform-server": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^5.3.1",
+    "@reactivex/rxjs": "^6.5.5",
+    "@types/moment-timezone": "^0.5.13",
+    "@types/pouchdb": "^6.4.0",
+    "@types/pouchdb-core": "^6.4.3",
+    "@types/uuid": "^3.4.9",
+    "bootstrap": "^4.4.1",
+    "buffer": "^5.6.0",
+    "chart.js": "^2.9.3",
+    "core-js": "^3.6.5",
     "dns": "^0.2.2",
-    "jquery": "^3.3.1",
-    "moment-timezone": "^0.5.16",
+    "jquery": "^3.5.0",
+    "moment-timezone": "^0.5.28",
     "ng2-charts": "^1.6.0",
-    "popper.js": "^1.14.3",
-    "rxjs": "^6.2.0",
-    "zone.js": "^0.8.26"
+    "popper.js": "^1.16.1",
+    "rxjs": "^6.5.5",
+    "uuid": "^7.0.3",
+    "zone.js": "^0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.6.8",
-    "@angular/compiler-cli": "^6.0.4",
-    "@angular/language-service": "^6.0.4",
-    "@types/jasmine": "^2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.2.1",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~1.7.1",
-    "karma-chrome-launcher": "~2.2.0",
+    "@angular-devkit/build-angular": "^0.800.6",
+    "@angular/cli": "^8.3.26",
+    "@angular/compiler-cli": "^8.2.14",
+    "@angular/language-service": "^8.2.14",
+    "@types/jasmine": "^3.5.10",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^13.13.2",
+    "codelyzer": "^5.2.2",
+    "jasmine-core": "^3.5.0",
+    "jasmine-spec-reporter": "^5.0.2",
+    "karma": "^5.0.2",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-cli": "^2.0.0",
     "karma-coverage-istanbul-reporter": "~1.4.2",
-    "karma-jasmine": "~1.1.1",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.3.0",
+    "karma-jasmine": "^3.1.1",
+    "karma-jasmine-html-reporter": "^1.5.3",
+    "node-sass": "^4.14.0",
+    "protractor": "^5.4.4",
     "ts-node": "~5.0.1",
-    "tslint": "~5.9.1",
-    "typescript": "~2.7.2"
+    "tslint": "^6.1.1",
+    "typescript": "^3.4.5",
+    "webpack": "^4.43.0"
   },
   "description": "This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 6.0.0.",
   "main": "index.js",
