@@ -1,9 +1,9 @@
-import { Component, OnInit } from "@angular/core";
+import { Component, OnInit } from '@angular/core';
 
 @Component({
-  selector: "app-bar-chart",
-  templateUrl: "./bar-chart.component.html",
-  styleUrls: ["./bar-chart.component.css"]
+  selector: 'app-bar-chart',
+  templateUrl: './bar-chart.component.html',
+  styleUrls: ['./bar-chart.component.css']
 })
 export class BarChartComponent {
   public barChartOptions: any = {
@@ -11,20 +11,20 @@ export class BarChartComponent {
     responsive: true
   };
   public barChartLabels: string[] = [
-    "2006",
-    "2007",
-    "2008",
-    "2009",
-    "2010",
-    "2011",
-    "2012"
+    '2006',
+    '2007',
+    '2008',
+    '2009',
+    '2010',
+    '2011',
+    '2012'
   ];
-  public barChartType: string = "bar";
-  public barChartLegend: boolean = true;
+  public barChartType = 'bar';
+  public barChartLegend = true;
 
   public barChartData: any[] = [
-    { data: [100, 59, 80, 81, 56, 55, 40], label: "Today" },
-    { data: [28, -48, 40, 200, 86, 27, 90], label: "Compare to yesterday" }
+    { data: [100, 59, 80, 81, 56, 55, 40], label: 'Today' },
+    { data: [28, -48, 40, 200, 86, 27, 90], label: 'Compare to yesterday' }
   ];
 
   // events
@@ -38,7 +38,7 @@ export class BarChartComponent {
 
   public randomize(): void {
     // Only Change 3 values
-    let data = [
+    const data = [
       Math.round(Math.random() * 100),
       59,
       80,
@@ -47,7 +47,7 @@ export class BarChartComponent {
       Math.random() * 100,
       40
     ];
-    let clone = JSON.parse(JSON.stringify(this.barChartData));
+    const clone = JSON.parse(JSON.stringify(this.barChartData));
     clone[0].data = data;
     this.barChartData = clone;
     /**
